apps/api/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-dotenv>=1.0.0
httpx>=0.26.0
